Create EditorWindowWorkspace base class
