+++ Workspaces/Common/Scripts/EditorWindowWorkspace.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: edeb17e963cfe0d4cbdbefb9957091fc
+timeCreated: 1492476891
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
